feat(\T_T/): i hate myself
make RNA, DNA & Protein implement Sequence and move around Aminoacid and
Codon
