+++ src/ctxs/fasta/application/to_aminoacids.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use crate::ctxs::{fasta::domain::fasta::Fasta, protein::domain::protein::Protein};
+use crate::ctxs::{fasta::domain::fasta::Fasta, protein::domain::Protein};
 
 pub fn to_aminoacids(file: PathBuf, ofile: Option<PathBuf>) -> Result<String, anyhow::Error> {
     let fasta: Fasta = crate::view::cat_f(&file)?;
