+++ src/ctxs/mod.rs
@@ -1,6 +1,5 @@
-pub mod aminoacid;
-pub mod codon;
 pub mod dna;
 pub mod fasta;
 pub mod protein;
+pub mod rna;
 pub mod shared;
