+++ src/ctxs/protein/mod.rs
@@ -1,2 +1,4 @@
+pub mod aminoacid;
+pub mod codon;
 pub mod domain;
 pub mod infrastructure;
